requirements.txt
discord.py
python-dotenv
supabase
rapidfuzz
numpy
matplotlib.pyplot
requests
pandas
io